requirements.txt
django-filebrowser<4
django-filter
django-grappelli<3
django-redis
django-rq
django>=3.2,<4
djangorestframework
drf-extensions

django-tabbed-admin
easy-thumbnails
pillow
psycopg2-binary
